fix next build by moving components to components dir
